Allow restarting failed workflows
This adds an attempt number that a client can use to restart a failed workflow.
It also restart failed workflows if a new version of the workflow is available.

To accomplish this, there is an attempt number and if the workflow is failed
and the client sends a higher attempt number, the steps in the workflow will be
restarted (after updating the arguments).