Remove support for optional parameters
Input parameters had the idea of being _required_. This was a copy of a design
in Shesmu for dealing with Cromwell and it can be replaced by something
simpler. Shesmu and Vidarr have the idea of optional values where a missing
value is encoded as JSON null. Cromwell has the idea of an optional value with
the same encoding, but also has parameters which should use a default, which
needed to be encoded as absent. Shesmu and Vidarr aped this "two ways to say
nothing" approach for compatibility, but this change unifies them. Shesmu and
Vidarr always encode optional values as null and the nulls are erased before
being passed to Cromwell.