Build jOOQ classes using Maven Docker
This modifies the build to run Postgres in Docker and have Flyway migrate the
schema to it and then generate the jOOQ objects from that database.

This trashes the existing Flyway migrations and since it is not currently
deployed and the workflow JSON data has changed, its fine to change the base
schema.