Allow deleting failed workflow runs
The submit endpoint allows rerunning a (possibly modified) failed workflow run.
This allows abandoning it entirely. It uses the same condition to determine if
the workflow can be removed.

This only applies to workflow runs that have not completed. Completed workflow
runs will be removed only using the unload mechanism not yet developed.