Change Cromwell engine parameter type parsing
Cromwell requires that the WDL options be a JSON object and the manual has this
as a map of names to types. The parser did not parse this as described in the
manual and would allow non-object types. This forces conversion to match the
manual and always provide an object.