Add `equals` and `hashCode` that ignore attempt
Shesmu uses the `equals` and `hashCode` methods to decide if two submission
requests are equivalent. However, it allows the user to increase the attempt
number and this doesn't really matter for equality. This adds a second pair of
methods to exclude the attempt number, i n the vein of
`String.equalsIgnoreCase`.