Allow reattempting unstarted workflows
Currently, a workflow has to be failed to allow a reattempt by the user.
However, if a workflow is waiting for resources, it also makes sense to allow
it be reattempted or deleted since it could be configured in a broken way where
it will never reach a failed state. This extends the existing reattempt checks
to include the `WAITING_FOR_RESOURCES` state and to stop attempting resource
checks if the workflow has been killed.