Fix contract bug in Cromwell output provisioner and workflow engine
The Vidarr interface requires that no real work be done in the `provision`,
`run`, or `recover` methods. The Cromwell output provisioner's `provision` and
`recover` started a job immediately and the Cromwell workflow engine's
`recover` might start a job immediately. This uses the scheduler to do these
operations only after the calling method is done.