Create an endpoint to access a single workflow
This endpoint allows checking for information on a single workflow. Unlike
`/api/workflows`, it always return a value, even if no versions are registered.
This is intended for use by the deploy script: if multiple variants of a
workflow are present, but not all variants (_e.g._, `bamqc_lane_level` and
`bamqc_call_ready`) are installed on different servers, it can push only the
appropriate versions.