Change version query generation to avoid SQL error
The version handling in the database is complicated. For each external key,
there are multiple versions and each version can have a different key (name)
and value. It is possible to have multiple entries with the same key though. In
the `VersionPolicy.NONE`, all of this information is discarded. In
`VersionPolicy.ALL`, it is aggregated into a map of lists. The generated query
for `ALL` has to do a subselect to get all the distinct keys that are available
and then another subselect to aggregate the values; `GROUP BY` does not work in
this context. `VersionPolicy.LATEST` provides only the most recent value. It
was using a `GROUP BY` operation, however, this failed when there were multiple
values for a single key.

This change replaces the logic to use the logic of `ALL` for `LATEST` with a
different aggregation operation.