Remove version export restriction
The `/api/workflows` endpoint only exported the latest version of a workflow to
reduce the amount of clutter in Shesmu. This is a bad idea for multiple
reasons:

- migration creates new old workflows, pushing the current workflow off the list
- the author of a new workflow might not want to immediately move all olives to the new version

So, I am removing this restriction and exporting all workflows. If there needs
to be export control per version, this can be added later.