Add some API integration tests (#99)
IMPORTANT NOTE: With this PR, the non-test dependencies are copied during the Vidarr build, and the step `mvn dependency:copy-dependencies` in your local build/deploy script needs to be removed. If it is not removed, the copy-dependencies step will copy ALL dependencies, including test deps, there will be module collisions, and Vidarr will fail to start.