Improve PAM input label
  * If the PAM plugin is configured, each workflow run submission
needs to contain this label-value pair. The initial
"required_services" label makes it sound like the value should be
a service that needs to be up, but alerts instead tend to fire if
something is down. This attempts to make it clearer what sort of
value (AutoInhibit alert labels) should be provided.