Max-In-Flight bugfixes (#138)
Adds workflow run id to inflight set on request() at line 48 to mirror line 36.
hander.launched and handler.resubmit take Runnables as a parameter. ConsumableResourceChecker implements Runnable and the intention was to pass one to both methods, however this was overlooked on resubmit's call.
inline comments to clarify the distinct codepaths for new vs reattempted workflow runs.
In extraordinarily rare circumstances, concurrent requests for the inflight
size would get the same number and could potentially push the inflight count
past max in flight to max in flight + # of cpus. this is not an emergency
but it could look like a bug and waste developers time so prevent it by synchronizing request methods.