GP-3269 Fix issue where Vidarr is resolving incorrect file paths
This metadata cache is only used when transitioning a workflow run to active, and we don't use Vidarr federation yet so it doesn't save that much time. If removing the cache proves to be an issue, we can reimplement it and properly remove items from the cache when they are unloaded.