Only count runs with greater priority against limit
Because of how the Comparator is implemented,
headSet includes all the runs with the same priority and hash
closer than 0. This is a problem because this means 100 jobs with
the same priority as you but a 'better' hash will keep you from
ever running. However we need that Comparator for functions like
add and remove, not considering the hash id means only 1 job of
each priority may enter the Set.

Switch compute logic to filter the set of active(/ly waiting) runs
to only jobs with a priority greater than you. In the context of
the max in flight scorer, this means only get the jobs set to
MAX_VALUE. If there is room then any job may launch regardless of
the 'worth' of its hash.