Split release into hard (complete) and soft (putItBack)
For aggregate scoring, it becomes a race condition problem
if we forget every run entirely when a different scorer
is unable to adequately score the job. Reset back to its
original priority rather than forget it entirely. We can
forget the job when the job has truly finished.

Also a bunch of formatting, sorry