Fix overzealous queue resetting
If we try to launch a wfr and it gets rejected by the Resource
Optimizing Priority Scorer because workflow max in flight is already
maxed out, we were still resetting the priority queue. This means
the queue may have only been set to 0 for a sliver of time.
Messy solution because of lack of workflow max in flight info at this
point in the code, but detect by symptom who approved and disapproved
the wfr and if it wasn't us, then reset the queue, otherwise trust
our original verdict and preserve the queue.